move NewTSOServiceCommand back to main
